Replace Travis with GitHub Actions (#101)
Currently, we use Travis CI for this repo. This change will move us to GitHub
Actions.

These are some of the things we need to do so we have parity with Travis.

- [x] build
- [x] pass tests
- [x] upload coverage
- [x] release*

goveralls doesn't support github actions.
https://docs.coveralls.io/supported-ci-services

codecov does though! https://docs.codecov.io/docs/supported-ci-providers

This PR switches this repo from coveralls to codecov for coverage reports.

*I need this PR to merge so I can test the release workflow.

This currently tests with Go 1.13 and 1.12.9. In addition, it tests with Linux,
macOS, and Windows.

+++ scripts/cov.sh
@@ -1,17 +1,32 @@
-#!/bin/bash -e
+#!/usr/bin/env bash
+set -ex
+
 # Run from directory above via ./scripts/cov.sh
+
+if [[ "$OSTYPE" != "linux-gnu" ]]; then
+	echo "$OSTYPE detected, skipping upload overage"
+	exit 0
+fi
+
+if [[ "$(go version)" != *"1.13"* ]]; then
+	echo "$(go version) detected, skipping upload overage"
+	exit 0
+fi
+
+echo "--- creating coverage report ---"
 
 rm -rf /tmp/cov
 mkdir -p /tmp/cov
 go test -v -covermode=atomic -coverprofile=/tmp/cov/collector.out ./collector
 go test -v -covermode=atomic -coverprofile=/tmp/cov/exporter.out ./exporter
-gocovmerge /tmp/cov/*.out > /tmp/acc.out
+$GOPATH/bin/gocovmerge /tmp/cov/*.out > coverage.txt
 rm -rf /tmp/cov
 
-# If we have an arg, assume travis run and push to coveralls. Otherwise launch browser results
+# If we have an arg, assume travis run and push to coveralls. Otherwise launch
+# browser results
 if [[ -n $1 ]]; then
-    $HOME/gopath/bin/goveralls -coverprofile=/tmp/acc.out -service travis-ci
-    rm -rf /tmp/acc.out
+	echo "--- uploading test coverage ---"
+	bash <(curl -s https://codecov.io/bash)
 else
     go tool cover -html=/tmp/acc.out
 fi
